requirements.txt
pandas>=1.5.0
numpy
statsmodels
linearmodels
openpyxl
pyarrow
fastparquet
pytest
scipy
anthropic>=0.25.0
python-dotenv>=1.0.0
